requirements.txt
# V3.1 "Dark Forest" Whale Hunter - Python Dependencies
# Install with: pip install -r requirements.txt

# Core
python-dotenv>=1.0.0
pyyaml>=6.0.1

# Async HTTP
aiohttp>=3.9.0
aiofiles>=23.2.1

# Database
# SQLite is built into Python - no extra package needed
# For Neo4j (Phase 3+):
# neo4j>=5.15.0

# Data Processing
pandas>=2.1.0
numpy>=1.26.0

# Solana
solders>=0.20.0
solana>=0.32.0
base58>=2.1.1

# API Clients
websockets>=12.0
httpx>=0.26.0

# Rate Limiting
aiohttp-retry>=2.8.3

# Logging & Monitoring
rich>=13.7.0
loguru>=0.7.2

# Testing
pytest>=7.4.0
pytest-asyncio>=0.23.0
pytest-cov>=4.1.0

# Development
black>=23.12.0
flake8>=7.0.0
mypy>=1.8.0
isort>=5.13.0

# Optional: Type stubs
types-pyyaml>=6.0.12
types-aiofiles>=23.2.0

# Optional: Performance profiling
# py-spy>=0.3.14
# memory-profiler>=0.61.0

# Phase 3+ Optional Dependencies:
# neo4j>=5.15.0          # Graph database
# redis>=5.0.0           # Caching
# fastapi>=0.109.0       # Webhook receiver
# uvicorn>=0.27.0        # ASGI server
